requirements.txt
numpy
pygame
PyOpenGL
gymnasium
pettingzoo
supersuit
stable-baselines3
tensorboard
